Give every game card the same shape, chrome and title
The /games index was five hand-written card blocks wired to a
grid-template-areas map that named each game three times over, with two
photo tiles and three SVG tiles at different aspect ratios per
breakpoint. Adding a game meant editing CSS in three places.

Now the manifest in games.js is the only list, one GameCard renders
every tile at a portrait 1 / 2, and the grid is a single auto-fit rule
that names no game. The title is sized in container query units off the
card rather than viewport media queries, because with a fluid column
count a wider window can mean a narrower card.

The three SVG artworks move verbatim into _artwork/; the two jpgs are
wrapped there unchanged so #119 has one file each to replace.

Closes #120
