+++ CLAUDE.md
@@ -55,6 +55,7 @@ npx prisma db push # apply schema changes
 - The same card is embedded on this site's `/github` page as `<img src="/api/github/card">` (no key — the route allows same-origin requests).
 - The card route is gated: requests must come from GitHub's Camo image proxy (`github-camo` User-Agent — README `<img>`s are proxied, never fetched by the viewer's browser, so there is no `Referer` to match on) **and** carry the `GITHUB_CARD_KEY` secret as `?key=`. Same-origin requests (the `/github` page) bypass both checks via `Sec-Fetch-Site`, falling back to a `Referer` host match. It is rate-limited in-memory (`src/lib/rate-limit.js`, 30 req/min/IP by default). Upstream GitHub GraphQL results are memoized with `unstable_cache` (60s) in `src/lib/github/api.js` (`getOverview`/`getStreak`) so the PAT is called at most once/minute regardless of request volume — this is what keeps `GITHUB_STATS_TOKEN` from being rate-limited. The route's response `Cache-Control` (consumed by GitHub's Camo proxy) is also 60s, so the rendered streak/stats stay near-live (subject to Camo's own internal minimum cache TTL).
 - The card SVG must not use CSS animations that start from `opacity: 0`. It is displayed inside an `<img>` (both on `/github` and on the README), and Chrome rasterizes an SVG image at animation time 0 without advancing it — a fade-in leaves the animated content permanently invisible.
+- The `/games` index is data-driven: `src/app/(pages)/games/games.js` is the manifest (`{ slug, title, href, Artwork }` per game, in display order) and `page.jsx` is one `.map` over it. Every tile is the same `GameCard` from `_components/GameCard.jsx` — the **card** owns the portrait `1 / 2` aspect ratio, the chrome and the title overlay, so an artwork never sets its own shape; artwork is authored as a square `viewBox="0 0 300 300"` (or an `<Image fill>`) in one module per game under `_artwork/`. The grid is a single `repeat(auto-fit, minmax(min(9rem, 100%), 1fr))` rule and the title is sized in container query units off the card, so adding a game means adding a manifest entry and an artwork file and touching no CSS.
 - `/games/tic-tac-overflow` is a two-player tic-tac-toe variant in `src/app/(pages)/games/tic-tac-overflow/`. Each player keeps at most three marks (`MAX_MARKS`); placing a fourth clears **that player's** oldest one, tracked as a per-player queue in `state.history`. `TicTacOverflow.jsx` is the mode picker: **same device** is the original hotseat game (its own `useReducer`, zero network calls — the realtime core is not in that branch at all) and **online** is the same board driven by a room. `?room=CODE` deep-links into a room; the effect that reads it runs after mount because the page is prerendered.
 - Both modes fill the *same* `context/index.jsx` Context — `{ state, dispatch, canPlay, canReset, online }` — so Cell, Status, Toolbar and Play Again are shared verbatim. `OnlineGame.jsx` fills it from `useRoom` instead of `useReducer` and translates the existing action creators into `send()`. `online` is null hotseat and carries the per-player view (your mark, opponent, room code, connection) when networked; `canPlay` closes the board on the opponent's turn and `canReset` makes "play again" a post-win rematch rather than a mid-game wipe.
 - The multiplayer definition is `multiplayer.js` (id `tto`), registered in `src/lib/realtime/registry.js`. It **calls** the hotseat `placeMark` reducer rather than restating the expiry rule — one implementation, server-authoritative. It is named `multiplayer.js` and not the `game.js` the core's README suggests because `Game.jsx` sits beside it and webpack treats two modules differing only in case as a collision on Windows/macOS. Its `createState`/`PLAY AGAIN` go through `createBoard(first)`, which is also what seeds hotseat `defaultState` — and `PLAY AGAIN` passes the winner, which is how the winner opens the next game online while hotseat always opens with X.
